+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - NV centers - `BasicElectronicNVElement` parameters accept physics-motivated values (!551)
 - Documentation - Sphinx build now compatible with qcodes==0.36.0 (!552)
 - Compilation - Can optionally provide a `QuantumDevice` to `QuantifyCompiler`. This will be used as default `CompilationConfig` in `QuantifyCompiler.compile()` (!535)
+- QuantumDevice - `BasicTransmonElement` can now be serialized to json string and deserialized via ``__getstate__/__init__`` (!510)
 
 ## 0.10.1 (2022-12-20)